+++ ui/media/js/main.js
@@ -59,14 +59,6 @@ let serverStatusColor = document.querySelector('#server-status-color')
 let serverStatusMsg = document.querySelector('#server-status-msg')
 
 
-document.querySelector('.drawing-board-control-navigation-back').innerHTML = '<i class="fa-solid fa-rotate-left"></i>'
-document.querySelector('.drawing-board-control-navigation-forward').innerHTML = '<i class="fa-solid fa-rotate-right"></i>'
-
-let maskResetButton = document.querySelector('.drawing-board-control-navigation-reset')
-maskResetButton.innerHTML = 'Clear'
-maskResetButton.style.fontWeight = 'normal'
-maskResetButton.style.fontSize = '10pt'
-
 let serverState = {'status': 'Offline', 'time': Date.now()}
 let bellPending = false
 
@@ -138,6 +130,35 @@ function isServerAvailable() {
     }
 }
 
+// shiftOrConfirm(e, prompt, fn)
+//   e      : MouseEvent
+//   prompt : Text to be shown as prompt. Should be a question to which "yes" is a good answer.
+//   fn     : function to be called if the user confirms the dialog or has the shift key pressed
+//
+// If the user had the shift key pressed while clicking, the function fn will be executed.
+// If the setting "confirm_dangerous_actions" in the system settings is disabled, the function 
+// fn will be executed.
+// Otherwise, a confirmation dialog is shown. If the user confirms, the function fn will also
+// be executed.
+function shiftOrConfirm(e, prompt, fn) {
+    e.stopPropagation()
+    if (e.shiftKey || !confirmDangerousActionsField.checked) {
+         fn(e)
+    } else {
+        $.confirm({
+            theme: 'modern',
+            title: prompt,
+            useBootstrap: false,
+            animateFromElement: false,
+            content: '<small>Tip: To skip this dialog, use shift-click or disable the "Confirm dangerous actions" setting in the Settings tab.</small>',
+            buttons: {
+                yes: () => { fn(e) },
+                cancel: () => {}
+            }
+        }); 
+    }
+}
+
 function logMsg(msg, level, outputMsg) {
     if (outputMsg.hasChildNodes()) {
         outputMsg.appendChild(document.createElement('br'))
@@ -168,34 +189,6 @@ function playSound() {
             console.warn("browser blocked autoplay")
         })
     }
-}
-function setSystemInfo(devices) {
-    let cpu = devices.all.cpu.name
-    let allGPUs = Object.keys(devices.all).filter(d => d != 'cpu')
-    let activeGPUs = Object.keys(devices.active)
-
-    function ID_TO_TEXT(d) {
-        let info = devices.all[d]
-        if ("mem_free" in info && "mem_total" in info) {
-            return `${info.name} <small>(${d}) (${info.mem_free.toFixed(1)}Gb free / ${info.mem_total.toFixed(1)} Gb total)</small>`
-        } else {
-            return `${info.name} <small>(${d}) (no memory info)</small>`
-        }
-    }
-
-    allGPUs = allGPUs.map(ID_TO_TEXT)
-    activeGPUs = activeGPUs.map(ID_TO_TEXT)
-
-    let systemInfo = `
-    <table>
-        <tr><td><label>Processor:</label></td><td class="value">${cpu}</td></tr>
-        <tr><td><label>Compatible Graphics Cards (all):</label></td><td class="value">${allGPUs.join('</br>')}</td></tr>
-        <tr><td></td><td>&nbsp;</td></tr>
-        <tr><td><label>Used for rendering 🔥:</label></td><td class="value">${activeGPUs.join('</br>')}</td></tr>
-    </table>`
-
-    let systemInfoEl = document.querySelector('#system-info')
-    systemInfoEl.innerHTML = systemInfo
 }
 
 async function healthCheck() {
@@ -231,7 +224,7 @@ async function healthCheck() {
                 break
         }
         if (serverState.devices) {
-            setSystemInfo(serverState.devices)
+            setDeviceInfo(serverState.devices)
         }
         serverState.time = Date.now()
     } catch (e) {
@@ -334,11 +327,7 @@ function onUseAsInputClick(req, img) {
     initImageSelector.value = null
     initImagePreview.src = imgData
 
-    initImagePreviewContainer.style.display = 'block'
-    inpaintingEditorContainer.style.display = 'none'
-    promptStrengthContainer.style.display = 'table-row'
     maskSetting.checked = false
-    samplerSelectionContainer.style.display = 'none'
 }
 
 function onDownloadImageClick(req, img) {
@@ -695,6 +684,12 @@ async function checkTasks() {
 
     const genSeeds = Boolean(typeof task.reqBody.seed !== 'number' || (task.reqBody.seed === task.seed && task.numOutputsTotal > 1))
     const startSeed = task.reqBody.seed || task.seed
+    
+    // Update the seed *before* starting the processing so it's retained if user stops the task
+    if (randomSeedField.checked) {
+        seedField.value = task.seed
+    }
+    
     for (let i = 0; i < task.batchCount; i++) {
         let newTask = task
         if (task.batchCount > 1) {
@@ -739,10 +734,6 @@ async function checkTasks() {
         if (task.outputMsg.innerText.toLowerCase().indexOf('error') === -1) {
             task.outputMsg.innerText = 'Task ended after ' + time + ' seconds'
         }
-    }
-
-    if (randomSeedField.checked) {
-        seedField.value = task.seed
     }
 
     currentTask = null
@@ -802,7 +793,7 @@ function getCurrentUserRequest() {
         //     newTask.reqBody.mask = maskImagePreview.src
         // }
         if (maskSetting.checked) {
-            newTask.reqBody.mask = inpaintingEditor.getImg()
+            newTask.reqBody.mask = imageInpainter.getImg()
         }
         newTask.reqBody.sampler = 'ddim'
     } else {
@@ -823,16 +814,26 @@ function getCurrentUserRequest() {
 function makeImage() {
     if (!isServerAvailable()) {
         alert('The server is not available.')
-        return
-    }
-    const taskTemplate = getCurrentUserRequest()
-    const newTaskRequests = []
-    getPrompts().forEach((prompt) => newTaskRequests.push(Object.assign({}, taskTemplate, {
-        reqBody: Object.assign({ prompt: prompt }, taskTemplate.reqBody)
-    })))
-    newTaskRequests.forEach(createTask)
+    } else if (!randomSeedField.checked && seedField.value == '') {
+        alert('The "Seed" field must not be empty.')
+    } else if (numOutputsTotalField.value == '') {
+        alert('The "Number of Images" field must not be empty.')
+    } else if (numOutputsParallelField.value == '') {
+        alert('The "Number of parallel Images" field must not be empty.')
+    } else if (numInferenceStepsField.value == '') {
+        alert('The "Inference Steps" field must not be empty.')
+    } else if (guidanceScaleField.value == '') {
+        alert('The Guidance Scale field must not be empty.')
+    } else {
+        const taskTemplate = getCurrentUserRequest()
+        const newTaskRequests = []
+        getPrompts().forEach((prompt) => newTaskRequests.push(Object.assign({}, taskTemplate, {
+            reqBody: Object.assign({ prompt: prompt }, taskTemplate.reqBody)
+        })))
+        newTaskRequests.forEach(createTask)
 
-    initialText.style.display = 'none'
+        initialText.style.display = 'none'
+    }
 }
 
 function createTask(task) {
@@ -877,24 +878,26 @@ function createTask(task) {
     task['progressBar'] = taskEntry.querySelector('.progress-bar')
     task['stopTask'] = taskEntry.querySelector('.stopTask')
 
-    task['stopTask'].addEventListener('click', async function(e) {
-        e.stopPropagation()
-        if (task['isProcessing']) {
-            task.isProcessing = false
-            task.progressBar.classList.remove("active")
-            try {
-                let res = await fetch('/image/stop?session_id=' + sessionId)
-            } catch (e) {
-                console.log(e)
-            }
-        } else {
-            let idx = taskQueue.indexOf(task)
-            if (idx >= 0) {
-                taskQueue.splice(idx, 1)
-            }
+    task['stopTask'].addEventListener('click', (e) => {
+        let question = (task['isProcessing'] ? "Stop this task?" : "Remove this task?")
+        shiftOrConfirm(e, question, async function(e) {
+            if (task['isProcessing']) {
+                task.isProcessing = false
+                task.progressBar.classList.remove("active")
+                try {
+                    let res = await fetch('/image/stop?session_id=' + sessionId)
+                } catch (e) {
+                    console.log(e)
+                }
+            } else {
+                let idx = taskQueue.indexOf(task)
+                if (idx >= 0) {
+                    taskQueue.splice(idx, 1)
+                }
 
-            taskEntry.remove()
-        }
+                removeTask(taskEntry)
+            }
+        })
     })
 
     task['useSettings'] = taskEntry.querySelector('.useSettings')
@@ -923,11 +926,12 @@ function getPrompts() {
     prompts = prompts.map(prompt => prompt.trim())
     prompts = prompts.filter(prompt => prompt !== '')
 
-    if (activeTags.length > 0) {
-	const promptTags = activeTags.map(x => x.name).join(", ")
-	prompts = prompts.map((prompt) => `${prompt}, ${promptTags}`)
+    const newTags = activeTags.filter(tag => tag.inactive === undefined || tag.inactive === false)
+    if (newTags.length > 0) {
+        const promptTags = newTags.map(x => x.name).join(", ")
+        prompts = prompts.map((prompt) => `${prompt}, ${promptTags}`)
     }
-	
+
     let promptsToMake = applySetOperator(prompts)
     promptsToMake = applyPermuteOperator(promptsToMake)
 
@@ -1038,21 +1042,25 @@ async function stopAllTasks() {
     }
 }
 
-clearAllPreviewsBtn.addEventListener('click', async function() {
+function removeTask(taskToRemove) {
+    taskToRemove.remove()
+
+    if (document.querySelector('.imageTaskContainer') === null) {
+        previewTools.style.display = 'none'
+        initialText.style.display = 'block'
+    }
+}
+
+clearAllPreviewsBtn.addEventListener('click', (e) => { shiftOrConfirm(e, "Clear all the results and tasks in this window?", async function() {
     await stopAllTasks()
 
     let taskEntries = document.querySelectorAll('.imageTaskContainer')
-    taskEntries.forEach(task => {
-        task.remove()
-    })
+    taskEntries.forEach(removeTask)
+})})
 
-    previewTools.style.display = 'none'
-    initialText.style.display = 'block'
-})
-
-stopImageBtn.addEventListener('click', async function() {
+stopImageBtn.addEventListener('click', (e) => { shiftOrConfirm(e, "Stop all the tasks?", async function(e) {
     await stopAllTasks()
-})
+})})
 
 widthField.addEventListener('change', onDimensionChange)
 heightField.addEventListener('change', onDimensionChange)
@@ -1073,13 +1081,14 @@ numOutputsTotalField.addEventListener('change', renameMakeImageButton)
 numOutputsParallelField.addEventListener('change', renameMakeImageButton)
 
 function onDimensionChange() {
-    if (!maskSetting.checked) {
-        return
-    }
     let widthValue = parseInt(widthField.value)
     let heightValue = parseInt(heightField.value)
-
-    resizeInpaintingEditor(widthValue, heightValue)
+    if (!initImagePreviewContainer.classList.contains("has-image")) {
+        imageEditor.setImage(null, widthValue, heightValue)
+    }
+    else {
+        imageInpainter.setImage(initImagePreview.src, widthValue, heightValue)
+    }
 }
 
 diskPathField.disabled = !saveToDiskField.checked
@@ -1192,7 +1201,7 @@ async function getModels() {
 function checkRandomSeed() {
     if (randomSeedField.checked) {
         seedField.disabled = true
-        seedField.value = "0"
+        //seedField.value = "0" // This causes the seed to be lost if the user changes their mind after toggling the checkbox
     } else {
         seedField.disabled = false
     }
@@ -1202,10 +1211,6 @@ checkRandomSeed()
 
 function showInitImagePreview() {
     if (initImageSelector.files.length === 0) {
-        initImagePreviewContainer.style.display = 'none'
-        // inpaintingEditorContainer.style.display = 'none'
-        promptStrengthContainer.style.display = 'none'
-        // maskSetting.style.display = 'none'
         return
     }
 
@@ -1213,13 +1218,7 @@ function showInitImagePreview() {
     let file = initImageSelector.files[0]
 
     reader.addEventListener('load', function(event) {
-        // console.log(file.name, reader.result)
         initImagePreview.src = reader.result
-        initImagePreviewContainer.style.display = 'block'
-        inpaintingEditorContainer.style.display = 'none'
-        promptStrengthContainer.style.display = 'table-row'
-        samplerSelectionContainer.style.display = 'none'
-        // maskSetting.checked = false
     })
 
     if (file) {
@@ -1230,34 +1229,25 @@ initImageSelector.addEventListener('change', showInitImagePreview)
 showInitImagePreview()
 
 initImagePreview.addEventListener('load', function() {
-    inpaintingEditorCanvasBackground.style.backgroundImage = "url('" + this.src + "')"
-    // maskSetting.style.display = 'block'
-    // inpaintingEditorContainer.style.display = 'block'
+    promptStrengthContainer.style.display = 'table-row'
+    initImagePreviewContainer.classList.add("has-image")
+
     initImageSizeBox.textContent = initImagePreview.naturalWidth + " x " + initImagePreview.naturalHeight
-    initImageSizeBox.style.display = 'block'
+    imageEditor.setImage(this.src, initImagePreview.naturalWidth, initImagePreview.naturalHeight)
+    imageInpainter.setImage(this.src, parseInt(widthField.value), parseInt(heightField.value))
 })
 
 initImageClearBtn.addEventListener('click', function() {
     initImageSelector.value = null
-    // maskImageSelector.value = null
-
     initImagePreview.src = ''
-    // maskImagePreview.src = ''
     maskSetting.checked = false
 
-    initImagePreviewContainer.style.display = 'none'
-    // inpaintingEditorContainer.style.display = 'none'
-    // maskImagePreviewContainer.style.display = 'none'
-
-    // maskSetting.style.display = 'none'
-
     promptStrengthContainer.style.display = 'none'
-    samplerSelectionContainer.style.display = 'table-row'
-    initImageSizeBox.style.display = 'none'
+    initImagePreviewContainer.classList.remove("has-image")
+    imageEditor.setImage(null, parseInt(widthField.value), parseInt(heightField.value))
 })
 
 maskSetting.addEventListener('click', function() {
-    inpaintingEditorContainer.style.display = (this.checked ? 'block' : 'none')
     onDimensionChange()
 })
 
@@ -1297,9 +1287,22 @@ document.querySelectorAll('.popup').forEach(popup => {
     }
 })
 
-var tabElements = [];
+var tabElements = []
+function selectTab(tab_id) {
+    let tabInfo = tabElements.find(t => t.tab.id == tab_id)
+    if (!tabInfo.tab.classList.contains("active")) {
+        tabElements.forEach(info => {
+            if (info.tab.classList.contains("active")) {
+                info.tab.classList.toggle("active")
+                info.content.classList.toggle("active")
+            }
+        })
+        tabInfo.tab.classList.toggle("active")
+        tabInfo.content.classList.toggle("active")
+    }
+}
 function linkTabContents(tab) {
-    var name = tab.id.replace("tab-", "");
+    var name = tab.id.replace("tab-", "")
     var content = document.getElementById(`tab-content-${name}`)
     tabElements.push({
         name: name,
@@ -1307,18 +1310,7 @@ function linkTabContents(tab) {
         content: content
     })
 
-    tab.addEventListener("click", event => {
-        if (!tab.classList.contains("active")) {
-            tabElements.forEach(tabInfo => {
-                if (tabInfo.tab.classList.contains("active")) {
-                    tabInfo.tab.classList.toggle("active")
-                    tabInfo.content.classList.toggle("active")
-                }
-            })
-            tab.classList.toggle("active")
-            content.classList.toggle("active")
-        }
-    })
+    tab.addEventListener("click", event => selectTab(tab.id))
 }
 
 document.querySelectorAll(".tab").forEach(linkTabContents)